Toki got the "Enkaigei".
Toki got the "Enkaigei".

+++ TohokuMouse2016/HAL/Adc.c
@@ -22,14 +22,14 @@ u16 AdcRead(u8 channel) {
 	while(S12AD.ADCSR.BIT.ADST);
 	
 	switch (channel) {
-		case  0 : value = S12AD.ADDR0; break;
-		case  1 : value = S12AD.ADDR1; break;
-		case  2 : value = S12AD.ADDR2; break;
-		case  6 : value = S12AD.ADDR6; break;
-		case  9 : value = S12AD.ADDR9; break;
-		case 10 : value = S12AD.ADDR10; break;
-		case 11 : value = S12AD.ADDR11; break;
-		case 12 : value = S12AD.ADDR12; break;
+		case  0 : value = (u16)(S12AD.ADDR0); break;
+		case  1 : value = (u16)(S12AD.ADDR1); break;
+		case  2 : value = (u16)(S12AD.ADDR2); break;
+		case  6 : value = (u16)(S12AD.ADDR6); break;
+		case  9 : value = (u16)(S12AD.ADDR9); break;
+		case 10 : value = (u16)(S12AD.ADDR10); break;
+		case 11 : value = (u16)(S12AD.ADDR11); break;
+		case 12 : value = (u16)(S12AD.ADDR12); break;
 
 		default : break;
 	}
